Fix rendering sprites in a different way.
This approach goes directly to creating an orthographic camera
with the desired projection.  Perhaps it better illustrates either the
bug within Projection::orthographic() or my misunderstanding of how it
is supposed to work.

I set the screen to black as well. The rest of the changes are just
rustfmt.

Compare to https://github.com/jlowry/amethyst/pull/1